Move data validation event header into export header
